+++ src/SearchLayout.js
@@ -1,6 +1,12 @@
 import React from 'react'
+import F from 'futil-js'
+import { observable } from 'mobx'
 import PropTypes from 'prop-types'
 import { observer } from 'mobx-react'
+
+let Context = React.createContext()
+
+export let useSearchLayout = () => React.useContext(Context)
 
 let styles = mode => ({
   display: 'grid',
@@ -11,9 +17,15 @@ let styles = mode => ({
     mode === 'basic' ? 'minmax(250px, 400px) minmax(0, 1fr)' : 'minmax(0, 1fr)',
 })
 
-let SearchLayout = ({ mode, style, ...props }) => (
-  <div style={{ ...styles(mode), ...style }} {...props} />
-)
+let SearchLayout = ({ style, mode, setMode, ...props }) => {
+  let [stateLayout] = React.useState(observable.box('basic'))
+  let layout = mode ? F.stateLens([mode, setMode]) : stateLayout
+  return (
+    <Context.Provider value={layout}>
+      <div style={{ ...styles(F.view(layout)), ...style }} {...props} />
+    </Context.Provider>
+  )
+}
 
 SearchLayout.propTypes = {
   mode: PropTypes.oneOf(['basic', 'builder', 'resultsOnly']),